backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
# torch
# torchvision
# torchaudio
sentence-transformers==2.7.0
transformers==4.41.0
faiss-cpu==1.13.0
numpy==1.26.4
nltk==3.8.1
pdfplumber==0.10.4
python-docx==0.8.11
aiofiles==23.2.1
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
# torch
# torchvision
# torchaudio
sentence-transformers==2.7.0
transformers==4.41.0
faiss-cpu==1.13.0
numpy==1.26.4
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
aiofiles==23.2.1
python-multipart==0.0.6
requests==2.31.0
groq==0.9.0
pytest==7.4.3
httpx==0.25.2
tqdm==4.66.0
pyyaml==6.0.1
accelerate==0.24.1
rouge-score==0.1.2
scikit-learn==1.3.2
google-generativeai